add update unit and change delete with confirmation

+++ app/Http/Controllers/UnitController.php
@@ -76,18 +76,6 @@ class UnitController extends Controller
     }
 
     /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Unit  $unit
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, Unit $unit)
-    {
-        //
-    }
-
-    /**
      * Remove the specified resource from storage.
      *
      * @param  \App\Models\Unit  $unit
@@ -98,9 +86,30 @@ class UnitController extends Controller
         //
     }
     public function delete(Request $request){
+        if(is_null($request->input('unit_id'))||empty($request->input('unit_id'))){
+            $request->flash('message', 'unit ID is required');
+            return redirect()->back();
+        }
         $id = $request->input('unit_id');
         Unit::destroy($id);
         Session::flash('message','Unit has been deletes');
+        return redirect()->back();
+    }
+    public function put(Request $request)
+    {
+        $request->validate([
+            'unit_code' => 'required',
+            'unit_id' => 'required',
+            'unit_name' => 'required'
+        ]);
+        $unitID = intval($request->input('unit_id'));
+
+        $unit = Unit::find($unitID);
+
+        $unit->unit_name = $request->input('unit_name');
+        $unit->unit_code = $request->input('unit_code');
+        $unit->save();
+        Session::flash('message', 'Unit ' . $unit->unit_name . ' has been updated');
         return redirect()->back();
     }
 }
